@@ -1,10 +1,10 @@
 name: sql_build_database
 description: A new Flutter package to create a database structure simply and quickly.
-version: 1.0.4
+version: 1.1.0
 homepage: https://github.com/DaniloCarSan/sql_build_database
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
